test(retrieval-attention): F-66 split prefill vs decode memory + F-67 fp16 selector features
F-67: cast contentNew to fp16 instead of fp32 in BatchedRetrievalAttentionIndex.update.
  Correctness preserved (F-34/F-42 multi-step match unchanged).
  perTokenFeatures buffer halves; total ~300 MB saved at 24K on 14B-1M.

F-66 enhanced: separate dense/RA peak memory for prefill vs single
decode step.

  [F-66] 14B-1M @ 24K
   dense_pref=17188 MB  ra_pref=21845 MB (+4657 MB / 27%)
   dense_dec =12677 MB  ra_dec =17880 MB (+5202 MB / 41%)

Surprising: RA decode-step shows MORE overhead than prefill (5.2 GB
vs 4.7 GB). Both phases have ~300 MB of selector state in the
RetrievalAttentionKVCache; the bulk (~4-5 GB) is in MLX intermediate
forward-pass buffers retained by the lazy graph.

Hypotheses for future investigation:
- MLXFast SDPA with mask=.array allocates score/attention buffers that
  persist longer than mask=.none/.causal fast paths
- The mask MLXArray itself (~96KB per layer per step) accumulates if
  not freed between layers
- per-Q-head broadcast may force materialization of intermediate
  attention probabilities across heads

Did NOT help (tested in this commit):
- Adding eval() barrier at end of update() for L>1 prefill — added
  250 MB instead of reducing
- fp16 storage of perTokenFeatures — saved ~300 MB but mask path
  overhead barely shifted

Acceptable headroom for ship: 32 GB Mac mini fits 14B-1M @ 24K + RA
with ~10 GB free.

+++ Libraries/MLXLMCommon/BatchedRetrievalAttentionIndex.swift
@@ -102,7 +102,10 @@ public final class BatchedRetrievalAttentionIndex {
         let newSeqLen = oldSeqLen + L
 
         // Project: [nKVHeads, L, dHead] @ [dHead, contentDim] = [nKVHeads, L, contentDim]
-        let contentNew = matmul(newKeys, WT).asType(.float32)
+        // F-67: store features as fp16 to halve perTokenFeatures memory.
+        // Score path upcasts on read; quality preserved (16-element dot
+        // products fit fp16 dynamic range cleanly).
+        let contentNew = matmul(newKeys, WT).asType(.float16)
 
         // Trig features only when lambdaPos > 0 — pure content (default
         // λ=0 ship config per F-46) ignores the trig half of the selector.
@@ -168,10 +171,12 @@ public final class BatchedRetrievalAttentionIndex {
                 isFine: false, featureDim: featureDim, multiToken: L > 1
             )
         }
-        // No explicit eval — in-place writes into pre-allocated buffers
-        // don't accumulate a deep graph chain like the old concat-based
-        // path did. The next consumer (asArray in topKBlockStartsCombined)
-        // will materialize the dependent block-buffer reads naturally.
+        // No explicit eval — in-place writes don't accumulate a deep
+        // graph chain. The next consumer materializes dependent reads
+        // naturally. (F-66 found 4.6 GB prefill overhead; adding eval
+        // here didn't reduce it — the bulk of the overhead is elsewhere
+        // in the forward-pass intermediate buffers, not in the index
+        // update path.)
     }
 
     /// In-place block-pool update onto a pre-allocated buffer.
